+++ lib/plugins/index.js
@@ -17,7 +17,7 @@ module.exports = [
   require('./slstats/slstats.js'),
   require('./plugin/plugin.js'),
   require('./plugin/install.js'),
-  require('./plugin/uninstall/uninstall.js'),
+  require('./plugin/uninstall.js'),
   require('./plugin/list.js'),
   require('./plugin/search.js'),
   require('./aws/configCredentials.js'),
