+++ lib/plugins/plugin/uninstall.js
@@ -5,8 +5,8 @@ const childProcess = BbPromise.promisifyAll(require('child_process'));
 const fse = require('fs-extra');
 const path = require('path');
 const _ = require('lodash');
-const yamlAstParser = require('../../../utils/yamlAstParser');
-const pluginUtils = require('../lib/utils');
+const yamlAstParser = require('../../utils/yamlAstParser');
+const pluginUtils = require('./lib/utils');
 
 class PluginUninstall {
   constructor(serverless, options) {
